fetch: change request method to GET for 301/302/303 status codes
Whenever a 301 or 302 is received and the original request method was
POST, the method is changed to GET and the request body and
content-length are dropped.
In case a 303 is received, the request method is always changed to GET,
no matter what the previous request method was. The body and
content-length are always dropped as well.

Furthermore, tests for this behavior are added.

Fix #19